+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file. See [standa
 ## 1.0.2 (2017-12-18)
 
 
+### Bug Fixes
+
+* **src:** adding entry points for direct access to the "Client" and "Server" modules ([ae964e3](https://github.com/Heliax44/udujs/commit/ae964e3))
+
+
 
 <a name="1.0.1"></a>
 ## 1.0.1 (2017-12-17)
+
+* Initial release.